russian_radarly/requirements.txt
boto3>=1.11.8
jsonschema==4.0.0
python-dotenv==0.10.5
requests==2.27.1
sentry-sdk==0.14.1
SQLAlchemy==1.4.25
